+++ src/checks/transport/mod.rs
@@ -3,5 +3,6 @@
 // Licensed under the MIT License <LICENSE-MIT or http://opensource.org/licenses/MIT>.
 // This file may not be copied, modified, or distributed except according to those terms.
 
+pub mod icmp;
 pub mod tcp;
 pub mod udp;