+++ src/errors/transport/mod.rs
@@ -7,6 +7,7 @@ use thiserror::Error;
 
 use crate::errors::transport::{tcp::TcpError, udp::UdpError};
 
+pub mod icmp;
 pub mod tcp;
 pub mod udp;