+++ src/parse/transport/protocols/mod.rs
@@ -7,6 +7,7 @@ use serde::Serialize;
 
 use crate::Transport;
 
+pub mod icmp;
 pub mod tcp;
 pub mod udp;